requirements.txt
click==8.1.3
dataclasses-json==0.5.7
dnspython==2.2.1
email-validator==1.3.0
flake8==5.0.4
Flask==2.2.2
Flask-Assets==2.0
Flask-Login==0.6.2
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytailwindcss==0.1.4
python-dotenv==0.21.0
typing-extensions==4.4.0
typing-inspect==0.8.0
webassets==2.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
